Add testing to constructor
